リポジトリ―およびentityの実装
