+++ app/src/main/java/com/api/EngineerCollabo/entities/Channel.java
@@ -9,10 +9,15 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.CascadeType;
+import java.util.ArrayList;
+import java.util.List;
+import jakarta.persistence.PrePersist;
 
 @Entity
-@Table(name = "directories")
-public class Directory {
+@Table(name = "channels")
+public class Channel {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
@@ -21,8 +26,16 @@ public class Directory {
     private String name;
 
     @ManyToOne()
-    @JoinColumn(name = "project_id", referencedColumnName = "id")
-    private Project project;
+    @JoinColumn(name = "user_id", referencedColumnName = "id")
+    private User user;
+
+    @ManyToOne()
+    @JoinColumn(name = "chat_room_id", referencedColumnName = "id")
+    private ChatRoom chatRoom;
+
+    @OneToMany(mappedBy = "channel", cascade = CascadeType.ALL)
+    private List<Message> messages = new ArrayList<>();
+
 
     public Integer getId() {
         return id;
