+++ app/src/main/java/com/api/EngineerCollabo/entities/Follower.java
@@ -9,20 +9,26 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.CascadeType;
+import java.util.ArrayList;
+import java.util.List;
+import jakarta.persistence.PrePersist;
 
 @Entity
-@Table(name = "user_notices")
-public class UserNotice {
+@Table(name = "Followers")
+public class Follower {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
 
-    @Column(name = "log", nullable = false)
-    private String log;
-
-    @ManyToOne()
+    @ManyToOne()    
     @JoinColumn(name = "user_id", referencedColumnName = "id")
     private User user;
+
+    @ManyToOne()    
+    @JoinColumn(name = "follower_id", referencedColumnName = "id")
+    private User follower;
 
     public Integer getId() {
         return id;
@@ -32,11 +38,19 @@ public class UserNotice {
         this.id = id;
     }
 
-    public String getLog() {
-        return log;
+    public User getUser() {
+        return user;
     }
 
-    public void setLog(String log) {
-        this.log = log;
+    public void setUser(User user) {
+        this.user = user;
+    }
+    
+    public User getFollower() {
+        return follower;
+    }
+    
+    public void setFollower(User follower) {
+        this.follower = follower;
     }
 }
