+++ app/src/main/java/com/api/EngineerCollabo/entities/Offer.java
@@ -9,6 +9,11 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.CascadeType;
+import java.util.ArrayList;
+import java.util.List;
+import jakarta.persistence.PrePersist;
 
 @Entity
 @Table(name = "offers")
@@ -25,8 +30,8 @@ public class Offer {
     private User scoutedUser;
 
     @ManyToOne()
-    @JoinColumn(name = "scout_user_id", referencedColumnName = "id")
-    private User scoutUser;
+    @JoinColumn(name = "user_id", referencedColumnName = "id")
+    private User user;
 
     public Integer getId() {
         return id;
