+++ app/src/main/java/com/api/EngineerCollabo/entities/Roll.java
@@ -9,30 +9,49 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.CascadeType;
+import java.util.ArrayList;
+import java.util.List;
+import jakarta.persistence.PrePersist;
 
 @Entity
-@Table(name = "members")
-public class Member {
+@Table(name = "rolls")
+public class Roll {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
+
+    @Column(name = "name", nullable = false)
+    private String name;
+
+    @Column(name = "count_log", nullable = false)
+    private Integer countLog;
 
     @ManyToOne()
     @JoinColumn(name = "user_id", referencedColumnName = "id")
     private User user;
 
-    @ManyToOne()
-    @JoinColumn(name = "project_id", referencedColumnName = "id")
-    private Project project;
-
-    // // 空のコンストラクタ
-    // public User() {}
-
-    // Getter、Setter
     public Integer getId() {
         return id;
     }
     public void setId(Integer id) {
         this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public Integer getCountLog() {
+        return countLog;
+    }
+    
+    public void setCountLog(Integer countLog) {
+        this.countLog = countLog;
     }
 }
