+++ app/src/main/java/com/api/EngineerCollabo/entities/Skill.java
@@ -9,6 +9,11 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.CascadeType;
+import java.util.ArrayList;
+import java.util.List;
+import jakarta.persistence.PrePersist;
 
 @Entity
 @Table(name = "skills")
@@ -24,8 +29,8 @@ public class Skill {
     private Integer countLog;
 
     @ManyToOne()
-    @JoinColumn(name = "member_id", referencedColumnName = "id")
-    private Member member;
+    @JoinColumn(name = "user_id", referencedColumnName = "id")
+    private User user;
 
     public Integer getId() {
         return id;
