+++ app/src/main/java/com/api/EngineerCollabo/entities/UserNotice.java
@@ -9,44 +9,39 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.CascadeType;
+import java.util.ArrayList;
+import java.util.List;
+import jakarta.persistence.PrePersist;
 
 @Entity
-@Table(name = "rolls")
-public class Roll {
+@Table(name = "user_notices")
+public class UserNotice {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
 
-    @Column(name = "name", nullable = false)
-    private String name;
-
-    @Column(name = "count_log", nullable = false)
-    private Integer countLog;
+    @Column(name = "log", nullable = false)
+    private String log;
 
     @ManyToOne()
-    @JoinColumn(name = "member_id", referencedColumnName = "id")
-    private Member member;
+    @JoinColumn(name = "user_id", referencedColumnName = "id")
+    private User user;
 
     public Integer getId() {
         return id;
     }
+
     public void setId(Integer id) {
         this.id = id;
     }
 
-    public String getName() {
-        return name;
+    public String getLog() {
+        return log;
     }
 
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public Integer getCountLog() {
-        return countLog;
-    }
-    
-    public void setCountLog(Integer countLog) {
-        this.countLog = countLog;
+    public void setLog(String log) {
+        this.log = log;
     }
 }
