--- app/src/main/java/com/api/EngineerCollabo/entity/Owner.java
@@ -1,41 +1,0 @@
-package com.api.EngineerCollabo;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Index;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.Table;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-
-@Entity
-@Table(name = "owners")
-public class Owner {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    @Column(name = "log", nullable = false)
-    private String log;
-
-    @ManyToOne()
-    @JoinColumn(name = "user_id", referencedColumnName = "id")
-    private User user;
-
-    @ManyToOne()
-    @JoinColumn(name = "project_id", referencedColumnName = "id")
-    private Project project;
-
-    // // 空のコンストラクタ
-    // public User() {}
-
-    // Getter、Setter
-    public Integer getId() {
-        return id;
-    }
-    public void setId(Integer id) {
-        this.id = id;
-    }
-}
