+++ app/src/main/java/com/api/EngineerCollabo/repositories/MessageRepository.java
@@ -1,7 +1,13 @@
 package com.api.EngineerCollabo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.List;
 
 public interface MessageRepository extends JpaRepository<Message, Integer> {
-    // カスタムのクエリメソッドが不要な場合は、このインターフェースだけで十分です
+
+    Message findById(int id);
+
+    List<Message> findByUser(User user);
+
+    List<Message> findByChannel(Channel channel);
 }
