+++ app/src/main/java/com/api/EngineerCollabo/repositories/OfferRepository.java
@@ -1,0 +1,15 @@
+package com.api.EngineerCollabo;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.List;
+
+public interface OfferRepository extends JpaRepository<Offer, Integer> {
+    
+    Offer findById(int id);
+
+    Offer findByMessage(String message);
+
+    List<Offer> findByUser(User user);
+
+    List<Offer> findByScoutedUser(User user);
+}
