+++ app/src/main/java/com/api/EngineerCollabo/repositories/ProjectRepository.java
@@ -1,0 +1,17 @@
+package com.api.EngineerCollabo;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.List;
+
+public interface ProjectRepository extends JpaRepository<Project, Integer> {
+    
+    Project findById(int id);
+
+    Project findByUsers(User user);
+
+    Project findByProjectNotices(ProjectNotice projectNotice);
+
+    Project findByDirectories(Directory directory);
+
+    Project findByFiles(File file);
+}
