+++ app/src/main/java/com/api/EngineerCollabo/repositories/RollRepository.java
@@ -1,7 +1,11 @@
 package com.api.EngineerCollabo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.List;
 
 public interface RollRepository extends JpaRepository<Roll, Integer> {
-    // カスタムのクエリメソッドが不要な場合は、このインターフェースだけで十分です
+
+    Roll findById(int id);
+
+    List<Roll> findByUser(User user);
 }
