+++ app/src/main/java/com/api/EngineerCollabo/repositories/UserNoticeRepository.java
@@ -1,7 +1,11 @@
 package com.api.EngineerCollabo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.List;
 
 public interface UserNoticeRepository extends JpaRepository<UserNotice, Integer> {
-    // カスタムのクエリメソッドが不要な場合は、このインターフェースだけで十分です
+
+    UserNotice findById(int id);
+
+    List<UserNotice> findByUser(User user);
 }
