--- app/src/main/java/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.api.EngineerCollabo;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.List;
-
-public interface UserRepository extends JpaRepository<User, Integer> {
-
-    User findById(int id);
-
-    List<User> findByName(String name);
-}
